modifying grid and simulation
In this commit, I completed update_scalar_concentration,
C is saved at the center at the beginning, at each step, c is interploated to the faces, then c is upated at the end of each RK4 substep. U is updated at update_particle and is the velocity field at the cell faces, CU is a multiplication of velocity and concentartion field a the cell face, and c is saved at center at the end of the sud step/steps.
Still have not modified write_info()